Simulate moved to WPF
