+++ WPF/SAM.Analytical.UI.WPF/Modify/SimulateCases.cs
@@ -58,7 +58,7 @@ namespace SAM.Analytical.UI.WPF
             using ProgressBarWindowManager progressBarWindowManager = new();
 
             progressBarWindowManager.Show("Running", "Running...");
-            Tas.Modify.RunWorkflow(analyticalModels, workflowSettings, directory, parallel, null, true);
+            Modify.RunWorkflow(analyticalModels, workflowSettings, directory, parallel, null, true);
             progressBarWindowManager.Close();
         }
     }
